requirements.txt
ultralytics==8.2.0
opencv-python
pillow==10.2.0